Rename snake_case for click_outside
