+++ packages/bbui/src/Form/Core/EnvDropdown.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
   import "@spectrum-css/textfield/dist/index-vars.css"
   import { createEventDispatcher, onMount } from "svelte"
-  import clickOutside from "../../Actions/click_outside"
+  import clickOutside from "../../Actions/clickOutside"
   import Divider from "../../Divider/Divider.svelte"
-  import type { EnvDropdownType } from "../../types"
   import Icon from "../../Icon/Icon.svelte"
+  import type { EnvDropdownType } from "../../types"
 
   export let value: string | number | undefined = undefined
   export let placeholder: string | undefined = undefined
